+++ CHANGELOG.rst
@@ -29,7 +29,8 @@ Unreleased
 * Extended test utils to compare dc buses, dc branches, FACTS, csc converters, and vsc converters.    
 * Added "set" method to PyParserMAT to follow C parser API.
 * Added shunt adjustment mode (continuous/discrete) and rounding capability.
-  
+* Exposed branch power flow Jacobian routine.
+
 Version 1.3.3
 -------------
 * Made Python wrapper its own repo.